Include configure-generated wx/setup.h if available
This is important under non-Apple systems too as this file defines
HAVE_UNISTD_H, for example, without which <unistd.h> doesn't get
included.

Note that we can't include it unconditionally, unfortunately, as the
include directory containing it is not in the compiler search path when
using MSVC, so for now only do it if a special symbol is predefined.
In the future it would be better to always set up the paths correctly.